Move Runtime documentation 2 levels up
This makes the github edit links all work for the Runtime repository. No
idea how it  worked beforehand. This creates a whole bunch of new
symlinks which have to be dealt with separately in the future. It's a
bit like the Runtime repo submoduling all the other projects inside
here. Also makes the building blocks _index file to be a symlink from
the Runtime repo.